Create elements.C
This program calculates the average of a given set of numbers. The user is prompted to enter the number of elements and then each number. The program stores the numbers in an array and calculates the sum. The average is then calculated by dividing the sum by the number of elements.